+++ Tests/Helper/Mocks/MockDecodableContainer.swift
@@ -12,6 +12,11 @@ import Hyperspace
 struct MockObject: Decodable {
     let title: String
     let subtitle: String
+    
+    init(title: String, subtitle: String) {
+        self.title = title
+        self.subtitle = subtitle
+    }
 }
 
 struct MockDate: Decodable {